requirements.txt
Flask
requests
python-dotenv
flasgger
transformers 
torch
huggingface-hub